feat: refactor CafeHome component into modular structure with Welcome, ActionButtons, and News sections
